move JWT data in security package
