+++ back/src/main/java/org/crunchytorch/coddy/security/filter/AuthorizationRequestFilter.java
@@ -1,7 +1,7 @@
 package org.crunchytorch.coddy.security.filter;
 
 import org.crunchytorch.coddy.security.service.JWTService;
-import org.crunchytorch.coddy.user.data.security.JWTPrincipal;
+import org.crunchytorch.coddy.security.data.JWTPrincipal;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
