+++ back/src/main/java/org/crunchytorch/coddy/security/service/JWTService.java
@@ -6,8 +6,8 @@ import org.apache.commons.lang.ArrayUtils;
 import org.apache.commons.lang.math.NumberUtils;
 import org.crunchytorch.coddy.application.utils.AppUtils;
 import org.crunchytorch.coddy.user.data.IUser;
-import org.crunchytorch.coddy.user.data.security.JWTPrincipal;
-import org.crunchytorch.coddy.user.data.security.JWTToken;
+import org.crunchytorch.coddy.security.data.JWTPrincipal;
+import org.crunchytorch.coddy.security.data.JWTToken;
 import org.crunchytorch.coddy.user.elasticsearch.entity.UserEntity;
 import org.crunchytorch.coddy.security.utils.SecurityUtils;
 import org.slf4j.Logger;
